shark_tank/requirements.txt
crewai
psycopg2-binary
sqlalchemy
python-dotenv